// DLP/ClientApp/src/api/models/company-branch-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KGH PROJECT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CompanyBranchDto
 */
export interface CompanyBranchDto {
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'id'?: string;
    /**
     * 
     * @type {number}
     * @memberof CompanyBranchDto
     */
    'ordinalNumber'?: number;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'branchOfficeName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'idNumber'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'address'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'email'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'contactPerson'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'contactPhone'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'place'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'municipalityId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'municipality'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'createdByUser'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'updatedAt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchDto
     */
    'updatedById'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CompanyBranchDto
     */
    'isDeleted'?: boolean;
}