// DLP/ClientApp/src/api/models/company-branch-dto-ordinal-paginated-list.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KGH PROJECT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { CompanyBranchDto } from './company-branch-dto';

/**
 * 
 * @export
 * @interface CompanyBranchDtoOrdinalPaginatedList
 */
export interface CompanyBranchDtoOrdinalPaginatedList {
    /**
     * 
     * @type {Array<CompanyBranchDto>}
     * @memberof CompanyBranchDtoOrdinalPaginatedList
     */
    'items'?: Array<CompanyBranchDto> | null;
    /**
     * 
     * @type {number}
     * @memberof CompanyBranchDtoOrdinalPaginatedList
     */
    'pageIndex'?: number;
    /**
     * 
     * @type {number}
     * @memberof CompanyBranchDtoOrdinalPaginatedList
     */
    'totalPages'?: number;
    /**
     * 
     * @type {number}
     * @memberof CompanyBranchDtoOrdinalPaginatedList
     */
    'totalCount'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof CompanyBranchDtoOrdinalPaginatedList
     */
    'hasPreviousPage'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CompanyBranchDtoOrdinalPaginatedList
     */
    'hasNextPage'?: boolean;
}